requirements.txt
web3==6.15.1
python-dotenv==1.0.1
py-solc-x==2.0.3
pandas==2.2.2
